tweak C data-types

+++ libpdf/buf.c
@@ -3,8 +3,8 @@
 #include <libpdf/buf.h>
 
 extern void pdf_buf_pack_8_4(uint8_t *out, uint8_t *in, size_t in_len) {
-  int i;
-  int j = 0;
+  size_t i;
+  size_t j = 0;
   for (i = 0; i < in_len; i++) {
     uint8_t v = in[i];
     out[j++] = v >> 4;
